Extract processed releases logic to the server
