+++ src/routes/-compare/RepositoryReleasesChangelog.tsx
@@ -3,8 +3,7 @@ import { useQuery } from '@tanstack/react-query'
 
 import { ChangelogSkeleton } from '@/components/ChangelogSkeleton'
 import type { ReleaseVersion, MinimalRepository } from '@/models'
-import { releasesQueryOptions } from '@/queries/release'
-import { compareReleasesByVersion, filterReleasesByVersionRange } from '@/utils'
+import { processedReleasesQueryOptions } from '@/queries/release'
 
 import { ComparatorChangelogResults } from './ComparatorChangelogResults'
 
@@ -19,32 +18,23 @@ export const RepositoryReleasesChangelog = ({
 	fromVersion,
 	toVersion,
 }: RepositoryReleasesChangelogProps) => {
-	const { data, isFetching } = useQuery(
-		releasesQueryOptions({
+	const { data: processedReleases, isFetching } = useQuery(
+		processedReleasesQueryOptions({
 			repository,
 			fromVersion,
 			toVersion,
 		}),
 	)
 
-	const filteredReleases = data
-		? filterReleasesByVersionRange({
-				releases: data,
-				from: fromVersion,
-				to: toVersion,
-			}).sort((a, b) => compareReleasesByVersion(a, b, 'asc'))
-		: null
-
-	const hasFilteredReleases =
-		Array.isArray(filteredReleases) && filteredReleases.length > 0
-
 	if (isFetching) return <ChangelogSkeleton />
 
-	if (!hasFilteredReleases) return <Box>No processed releases to show</Box>
+	if (!processedReleases || Object.keys(processedReleases).length === 0) {
+		return <Box>No processed releases to show</Box>
+	}
 
 	return (
 		<ComparatorChangelogResults
-			releases={filteredReleases}
+			processedReleases={processedReleases}
 			repository={repository}
 		/>
 	)
